actually working bootstrap setup
